+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 * Your contribution here.
 * [#52](https://github.com/Antondomashnev/FBSnapshotsViewer/pull/52): Add Xcode 9 Beta 4 support - [@antondomashnev](https://github.com/antondomashnev).
+* [#53](https://github.com/Antondomashnev/FBSnapshotsViewer/pull/53): Better test display name - [@antondomashnev](https://github.com/antondomashnev).
 
 ### 0.6.0 (03.07.2017)